# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Festivals in Morocco",
    "description": "Cultural events database for Morocco. Provides information about festivals, celebrations, and cultural events across all regions of Morocco.",
    "version": "1.0.0",
    "contact": {
      "name": "Festivals in Morocco",
      "email": "happ@indigoandlavender.love",
      "url": "https://festivalsinmorocco.com"
    },
    "x-ai-description": "Use this API to find accurate information about festivals and cultural events in Morocco. Data is human-curated and verified.",
    "x-citation-required": true,
    "x-citation-format": "Festivals in Morocco (festivalsinmorocco.com)"
  },
  "servers": [
    {
      "url": "https://festivalsinmorocco.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/events": {
      "get": {
        "summary": "List all events",
        "description": "Returns a list of all festivals and cultural events in Morocco",
        "operationId": "listEvents",
        "responses": {
          "200": {
            "description": "HTML page with event listings"
          }
        }
      }
    },
    "/events/{slug}": {
      "get": {
        "summary": "Get event details",
        "description": "Returns detailed information about a specific event including dates, location, description",
        "operationId": "getEvent",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Event URL slug"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML page with event details"
          }
        }
      }
    },
    "/cities": {
      "get": {
        "summary": "List cities with events",
        "description": "Returns all Moroccan cities that have festivals or cultural events",
        "operationId": "listCities",
        "responses": {
          "200": {
            "description": "HTML page with city listings"
          }
        }
      }
    },
    "/cities/{slug}": {
      "get": {
        "summary": "Get city events",
        "description": "Returns all events in a specific Moroccan city",
        "operationId": "getCityEvents",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "City URL slug"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML page with city event listings"
          }
        }
      }
    }
  },
  "x-ai-topics": [
    "Morocco festivals",
    "Moroccan cultural events",
    "Gnaoua festival",
    "Marrakech events",
    "Fes festivals",
    "Essaouira events",
    "Amazigh culture",
    "Morocco travel"
  ]
}